requirements.txt
discord.py
plyvel